fix(kernel): correctly scale distances
This is a bit of an embarassing bug, I hadn't actually scaled the
distances used in the kernel by the scale of the DEM. I don't think this
had a major effect though as all the angle calculations were still
relatively correct and so viewsheds were still mostly correct. The main
problem was that lines of sight weren't terminating due to the curvature
of the Earth.