feat: rotation kernel
This is mainly to improve memory alignment. But it also simplifies the
codebase by getting rid of all the code associated with the "band of
sight".

I setup proper ASCII based viewsheds for intuitive testing. This
helped find and fix a few bugs in the viewshed reconstructor.

Also I started exploring rotation interpolation. It appears that
bilinear leads to more accurate viewsheds, but there are still some
things to explore:
* Bilinear breaks the symmetry of the ASCII tests. Does this mean that
  bilinear is more or less accurate for elevations nearer the centre of
  rotation? Is it that bilinear makes larger viewsheds more accurate,
  but smaller viewsheds less accurate?
* Try other interpolation methods like spline 8 or 16.
* Explore only interpolating the elevations nearest the centre, say up
  to a radius of 90 points. 90 creates a square of "circumfrence" 360,
  so outside that radius interpolation doesn't counter act distortions.