fix: rotate line of sight, anti-rotate DEM
Previously I was rotating the DEM, which had the effect of
_anti_-rotating the line of sight. When debugging we're mostly looking
at IDs in the line of sight, we rarely, if ever, look at the DEM itself.
So it just makes more intuitive sense that a rotation always refers to
the line of sight.

More to the point, this is also the approach @ryan-berger is taking with
his kernel. And in order to be able to share tests, we need a unified
approach.

In theory this shouldn't change anything other than making debugging
easier. In reality it has slightly changed the benchmark viewshed and
the unit tested total surfaces. Floating points?